+++ lib/input/sqs.go
@@ -41,7 +41,21 @@ messages.
 By default Benthos will use a shared credentials file when connecting to AWS
 services. It's also possible to set them explicitly at the component level,
 allowing you to transfer data across accounts. You can find out more
-[in this document](../aws.md).`,
+[in this document](../aws.md).
+
+### Metadata
+
+This input adds the following metadata fields to each message:
+
+` + "```" + `
+- sqs_message_id
+- sqs_receipt_handle
+- sqs_approximate_receive_count
+- All message attributes
+` + "```" + `
+
+You can access these metadata fields using
+[function interpolation](../config_interpolation.md#metadata).`,
 	}
 }